Updated readme.
Most instances of array<mp_value@> were replaced with the likely more familiar mp_value@[]. The full syntax is no longer needed, it was only necessary when the handle type was const. Exceptions are notes on casting and conversion of value objects, where the syntax might actually make some difference and for correctness.
Fixed the signature of mp_decoder's constructor, changing string initial to string initial = "", which is more correct.
Moved the functions section to just below the enumerations section, putting it before classes. For the simplest of use cases, the decoder and encoder classes need not be touched.