Replaced all templated array specifications with bracket syntax in both readme and msgpack.nvgt, and some debug changes.
A new marker /// BEGIN DEBUG /// was added to mark the top of the debug section for script usage.
A basic script debugstrip.py was created to use this marker and searching for dbgout statements to convert a debug version of the library into a release one, to be used when the patch no longer cleanly reverts and a new one must be generated.
Updated debug.patch with the use of this script.
Updated README with rules about debug.patch.