Updated README.
Fixed explicit references to the constant MP_TYPE_MISMATCH_EXCEPTION, which previously didn't have the MP_ prefix when a namespace was used.
Added a constants section after enumerations which lists these exception constants and known extension types.
Documented the existence of these constants in the list of API provisions at the top.