Relaxed value equality comparison slightly and some other misc changes.
Equality comparison of value objects no longer requires they match exactly in format, and a best-effort check is made based on upcasting. This is done in case the serialized representation read from an incoming stream doesn't match this library's behavior exactly.
This allows floats to compare equal to doubles and integers of the same sign but different widths to compare equal to each other. Integers of different signs will always compare unequal, however.
A new signed property has been added to value objects, which determines if their stored format is signed or not. This cuts down on duplicated code in string conversion and equality checks for integers.
Added a new standard exception MP_LARGE_VALUE_EXCEPTION, which is now returned by all the constructors and format when a value too big to be represented is encountered. Previously these returned type-specific exceptions.
Reran astyle on the code, which preserved tabs this time, and removed the resulting spaces after throw statements.
Updated readme in accordance with the rest.